This commit improves code readability and reduces redundancy while improving a few page animations.
