+++ app/src/main/java/com/tpc/nudj/ui/components/LoadingOverlay.kt
@@ -1,0 +1,31 @@
+package com.tpc.nudj.ui.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.tpc.nudj.ui.theme.LocalAppColors
+
+@Composable
+@Preview(showBackground = true)
+fun LoadingScreenOverlay() {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(color = Color.Black.copy(alpha = 0.5f)),
+        contentAlignment = Alignment.Center
+    ) {
+        CircularProgressIndicator(
+            modifier = Modifier.size(50.dp),
+            color = LocalAppColors.current.appTitle,
+            strokeWidth = 6.dp
+        )
+    }
+}
